refactor(frontend): convert SessionSettingModal to SessionDrawer (BottomSheet)
- Replace ModalDialog with BottomSheet (auto height, #header slot)
- Rename component: SessionSettingModal → SessionDrawer
- Props: show → open, emits: update:show → close
- ChatInputBar: openSettingsModal → openSettingsDrawer
- Update all test mocks: ModalDialog → BottomSheet
- Update comments referencing SessionSettingModal
