--- web/src/components/__tests__/sessionSettingModal.test.ts
@@ -1,813 +1,0 @@
-import { describe, expect, it, vi, beforeEach } from 'vitest'
-import { mount } from '@vue/test-utils'
-import { nextTick, ref, defineComponent, h } from 'vue'
-import SessionSettingModal from '@/components/chat/SessionSettingModal.vue'
-import { useAgents } from '@/composables/useAgents'
-import { useSessionIdentity } from '@/composables/useSessionIdentity'
-import { apiPost } from '@/utils/api'
-import { patchAgentPref } from '@/composables/useSettingsConfig'
-
-// Mock ModalDialog to render slot content inline (skip Teleport).
-// Vue 3.5 + @vue/test-utils 2.4.11 broke teleport stubs — the stub renders
-// an empty element instead of slot content. Mocking ModalDialog avoids
-// Teleport entirely, keeping slot content in the component tree.
-vi.mock('@/components/common/ModalDialog.vue', () => ({
-  default: defineComponent({
-    props: { open: Boolean, title: String, zIndex: Number, fullHeight: Boolean },
-    emits: ['close'],
-    inheritAttrs: true,
-    template: `
-      <div class="modal-overlay">
-        <div class="modal-dialog">
-          <div class="modal-header"><slot name="header" /></div>
-          <div class="modal-body"><slot /></div>
-          <div class="modal-footer"><slot name="footer" /></div>
-          <slot name="after" />
-        </div>
-      </div>
-    `,
-  }),
-}))
-
-// Mock PopupMenu similarly — its <Teleport>+<Transition> breaks Vue 3.5
-// reactivity in test-utils. Render slot content inline without Teleport.
-vi.mock('@/components/common/PopupMenu.vue', () => ({
-  default: defineComponent({
-    props: { show: Boolean, targetElement: Object, maxWidth: Number, maxHeight: Number, menuItemsCount: Number },
-    emits: ['update:show'],
-    template: `
-      <div v-if="show" class="popup-menu-stub">
-        <slot />
-      </div>
-    `,
-  }),
-}))
-
-// Mock composables
-vi.mock('@/composables/useAgents', () => ({
-  useAgents: vi.fn(),
-  restoreOriginalModels: vi.fn(),
-  populateACPStateCache: vi.fn().mockResolvedValue(undefined),
-  invalidateACPStateCache: vi.fn(),
-}))
-vi.mock('@/composables/useSessionIdentity', () => ({
-  useSessionIdentity: vi.fn(),
-  clearModeState: vi.fn(),
-  clearCommandState: vi.fn(),
-  clearThinkingEffortState: vi.fn(),
-}))
-vi.mock('@/utils/api', () => ({
-  apiPost: vi.fn().mockResolvedValue({ models: [] }),
-}))
-vi.mock('vue-i18n', () => ({
-  useI18n: () => ({ t: (key: string) => key }),
-  createI18n: () => ({ global: { t: (key: string) => key } }),
-}))
-vi.mock('@/composables/useSettingsConfig', () => ({
-  patchAgentPref: vi.fn().mockResolvedValue(undefined),
-}))
-const mockToastShow = vi.fn()
-vi.mock('@/composables/useToast', () => ({
-  useToast: () => ({ show: mockToastShow }),
-}))
-vi.mock('@/composables/useLocale', () => ({
-  gt: (key: string) => key,
-}))
-
-const mockAgents = {
-  agents: ref([
-    {
-      id: 'claude',
-      name: 'Claude',
-      icon: '🤖',
-      backend: 'claude',
-      models: [
-        { id: 'claude-sonnet-4-6', name: 'Claude Sonnet 4.6', default: true },
-        { id: 'claude-opus-4-5', name: 'Claude Opus 4.5', default: false },
-        { id: 'claude-haiku-3-5', name: 'Claude Haiku 3.5', default: false },
-      ],
-      thinkingEffortLevels: ['low', 'medium', 'high', 'xhigh', 'max'],
-      preferredModel: 'claude-sonnet-4-6',
-      preferredThinkingEffort: '',
-      canRefreshModels: true,
-      acpCommand: 'npx -y @agentclientprotocol/claude-agent-acp@latest',
-      transport: 'acp-stdio',
-    },
-    {
-      id: 'kimi',
-      name: 'Kimi',
-      icon: '💎',
-      backend: 'kimi',
-      models: [],
-      thinkingEffortLevels: [],
-      preferredModel: '',
-      preferredThinkingEffort: '',
-      canRefreshModels: false,
-    },
-  ]),
-  getAgentModels: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return a?.models || []
-  }),
-  getAgentThinkingEffortLevels: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return a?.thinkingEffortLevels || []
-  }),
-  refreshAgentModels: vi.fn().mockResolvedValue(undefined),
-  updateAgentField: vi.fn(),
-  getDefaultModelId: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return a?.preferredModel || a?.models?.[0]?.id || ''
-  }),
-  getAgent: vi.fn((agentId: string) => {
-    return mockAgents.agents.value.find(a => a.id === agentId)
-  }),
-  canRefreshModels: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return !!a?.canRefreshModels
-  }),
-  supportsDualTransport: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return !!a?.acpCommand
-  }),
-  getAgentTransport: vi.fn((agentId: string) => {
-    const a = mockAgents.agents.value.find(a => a.id === agentId)
-    return a?.transport || 'cli'
-  }),
-}
-
-const mockIdentity = {
-  currentAgentId: ref('claude'),
-  currentModelId: ref('claude-sonnet-4-6'),
-  currentModelName: ref('Claude Sonnet 4.6'),
-  currentThinkingEffort: ref('high'),
-  currentTransport: ref('acp-stdio'),
-  availableThinkingEfforts: ref([]),
-  availableModes: ref([{ id: 'code', name: 'Code' }, { id: 'ask', name: 'Ask' }]),
-  currentModeId: ref('code'),
-  autoApprove: ref(false),
-  toggleAutoApprove: vi.fn(),
-}
-
-describe('SessionSettingModal', () => {
-  beforeEach(() => {
-    vi.mocked(useAgents).mockReturnValue(mockAgents as any)
-    vi.mocked(useSessionIdentity).mockReturnValue(mockIdentity as any)
-    vi.mocked(apiPost).mockResolvedValue({ models: [] })
-    vi.mocked(patchAgentPref).mockResolvedValue(undefined)
-    mockToastShow.mockClear()
-  })
-
-  function mountModal(props = {}) {
-    return mount(SessionSettingModal, {
-      props: { show: true, agentId: 'claude', ...props },
-    })
-  }
-
-  // --- Tab switching ---
-
-  it('renders model tab by default', () => {
-    const wrapper = mountModal()
-    expect(wrapper.find('.model-tab.active').text()).toContain('chat.modelSwitcher.title')
-  })
-
-  it('switches to thinking tab via initialTab prop', () => {
-    // Vue 3.5 + test-utils doesn't re-render on ref changes,
-    // so use initialTab prop to test the thinking tab state.
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const tabs = wrapper.findAll('.model-tab')
-    expect(tabs[1].classes()).toContain('active')
-  })
-
-  // --- Model list ---
-
-  it('renders model list for current agent', () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    expect(items.length).toBe(3)
-    expect(items[0].text()).toContain('Claude Sonnet 4.6')
-  })
-
-  it('highlights current session model', () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    expect(items[0].classes()).toContain('current')
-  })
-
-  it('shows default badge on preferred model', () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    // First model is preferred, should have default badge
-    expect(items[0].find('.default-badge').exists() || items[0].text().includes('默认')).toBe(true)
-  })
-
-  // --- Search ---
-  // Vue 3.5 + test-utils doesn't re-render on ref changes.
-  // Test search filtering via VM methods instead of DOM interaction.
-
-  it('filters models by search query', () => {
-    const wrapper = mountModal()
-    wrapper.vm._setSearchQuery('opus')
-    const filtered = wrapper.vm._getFilteredModels()
-    expect(filtered.length).toBe(1)
-    expect(filtered[0].name).toContain('Opus')
-  })
-
-  it('shows no results message when search has no matches', async () => {
-    const wrapper = mountModal()
-    const searchInput = wrapper.find('.model-search-input')
-    await searchInput.setValue('nonexistent')
-    await nextTick()
-
-    expect(wrapper.find('.model-empty').exists() || wrapper.text()).toBeTruthy()
-  })
-
-  it('filters models by id when search matches id but not name', () => {
-    const wrapper = mountModal()
-    wrapper.vm._setSearchQuery('haiku-3')
-    const filtered = wrapper.vm._getFilteredModels()
-    expect(filtered.length).toBe(1)
-    expect(filtered[0].name).toContain('Haiku')
-  })
-
-  it('shows no-search-results message when search yields nothing', () => {
-    const wrapper = mountModal()
-    wrapper.vm._setSearchQuery('xyz')
-    const filtered = wrapper.vm._getFilteredModels()
-    expect(filtered.length).toBe(0)
-  })
-
-  // --- Model selection (session-scoped) ---
-
-  it('emits switch-model when clicking a model', async () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    await items[1].trigger('click') // opus
-
-    expect(wrapper.emitted('switch-model')).toBeTruthy()
-    expect(wrapper.emitted('switch-model')![0][0]).toEqual({ id: 'claude-opus-4-5', name: 'Claude Opus 4.5', default: false })
-  })
-
-  it('closes modal after selecting a model', async () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    await items[1].trigger('click')
-
-    expect(wrapper.emitted('update:show')).toBeTruthy()
-    expect(wrapper.emitted('update:show')![0][0]).toBe(false)
-  })
-
-  // --- Thinking effort ---
-  // Use initialTab: 'thinking' to test thinking tab content.
-
-  it('renders thinking effort levels on thinking tab', () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    expect(items.length).toBe(5)
-  })
-
-  it('highlights current thinking effort', () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    // 'high' is current, find it
-    const highItem = items.find(i => i.text().includes('high'))
-    expect(highItem?.classes()).toContain('current')
-  })
-
-  it('emits switch-thinking-effort when clicking a level', async () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    // Click 'medium'
-    const mediumItem = items.find(i => i.text().includes('medium'))
-    await mediumItem?.trigger('click')
-
-    expect(wrapper.emitted('switch-thinking-effort')).toBeTruthy()
-  })
-
-  it('closes modal after selecting thinking effort', async () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    const mediumItem = items.find(i => i.text().includes('medium'))
-    await mediumItem?.trigger('click')
-
-    expect(wrapper.emitted('update:show')).toBeTruthy()
-    expect(wrapper.emitted('update:show')![0][0]).toBe(false)
-  })
-
-  it('shows default badge on default thinking effort level', () => {
-    const claudeAgent = mockAgents.agents.value.find(a => a.id === 'claude')!
-    claudeAgent.preferredThinkingEffort = 'high'
-
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    const highItem = items.find(i => i.text().includes('high'))
-    expect(highItem?.find('.default-badge').exists()).toBe(true)
-
-    claudeAgent.preferredThinkingEffort = '' // restore
-  })
-
-  // --- Refresh ---
-
-  it('has refresh button for agents that support model refresh', () => {
-    const wrapper = mountModal()
-    expect(wrapper.find('.refresh-btn').exists()).toBe(true)
-  })
-
-  it('hides refresh button for agents that do not support model refresh', () => {
-    const wrapper = mountModal({ agentId: 'kimi' })
-    expect(wrapper.find('.refresh-btn').exists()).toBe(false)
-  })
-
-  it('calls refresh API and updates agent models on success', async () => {
-    const newModels = [
-      { id: 'claude-new', name: 'Claude New', default: true },
-    ]
-    vi.mocked(apiPost).mockResolvedValue({ models: newModels })
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-    // Wait for async handler
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(apiPost).toHaveBeenCalledWith('/api/agents/claude/refresh-models', {})
-    expect(mockAgents.updateAgentField).toHaveBeenCalledWith('claude', 'models', newModels)
-    expect(mockToastShow).toHaveBeenCalledWith('chat.sessionSetting.refreshSuccess', expect.any(Object))
-  })
-
-  it('shows cliNotFound toast when CLI not found error', async () => {
-    vi.mocked(apiPost).mockRejectedValue({ msgKey: 'CLINotFound' })
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(mockToastShow).toHaveBeenCalledWith('chat.sessionSetting.cliNotFound', expect.any(Object))
-  })
-
-  it('shows discoveryNotSupported toast when model discovery not supported', async () => {
-    vi.mocked(apiPost).mockRejectedValue({ msgKey: 'ModelDiscoveryNotSupported' })
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(mockToastShow).toHaveBeenCalledWith('chat.sessionSetting.discoveryNotSupported', expect.any(Object))
-  })
-
-  it('shows generic refreshFailed toast on other errors', async () => {
-    vi.mocked(apiPost).mockRejectedValue(new Error('network error'))
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(mockToastShow).toHaveBeenCalledWith('chat.sessionSetting.refreshFailed', expect.any(Object))
-  })
-
-  it('disables refresh button while refreshing', async () => {
-    let resolveRefresh: (v: any) => void
-    vi.mocked(apiPost).mockReturnValue(new Promise(r => { resolveRefresh = r }))
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-
-    // Check the refreshing ref is true via VM
-    expect(wrapper.vm.refreshing).toBe(true)
-
-    resolveRefresh!({ models: [] })
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-  })
-
-  it('does not call API when already refreshing', async () => {
-    let resolveRefresh: (v: any) => void
-    vi.mocked(apiPost).mockReturnValue(new Promise(r => { resolveRefresh = r }))
-
-    const wrapper = mountModal()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-
-    // Try clicking again while still refreshing
-    vi.mocked(apiPost).mockClear()
-    await wrapper.find('.refresh-btn').trigger('click')
-    await nextTick()
-
-    expect(apiPost).not.toHaveBeenCalled()
-
-    resolveRefresh!({ models: [] })
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-  })
-
-  // --- Visual dividers ---
-
-  it('renders dividers between model items', () => {
-    const wrapper = mountModal()
-    const dividers = wrapper.findAll('.model-divider')
-    // 3 models = 2 dividers between them
-    expect(dividers.length).toBe(2)
-  })
-
-  // --- Set default button ---
-
-  it('has set-default star button on non-default models', () => {
-    const wrapper = mountModal()
-    const setDefaultBtns = wrapper.findAll('.set-default-btn')
-    // 3 models, 1 is default (no star btn), 2 have star btns
-    expect(setDefaultBtns.length).toBe(2)
-  })
-
-  it('calls patchAgentPref and updateAgentField when star button clicked', async () => {
-    const wrapper = mountModal()
-    const starBtns = wrapper.findAll('.set-default-btn')
-    await starBtns[0].trigger('click') // Click first star (opus)
-
-    expect(patchAgentPref).toHaveBeenCalledWith('claude', 'preferred_model', 'claude-opus-4-5')
-    expect(mockAgents.updateAgentField).toHaveBeenCalledWith('claude', 'preferredModel', 'claude-opus-4-5')
-  })
-
-  it('shows error toast when setDefaultModel fails', async () => {
-    vi.mocked(patchAgentPref).mockRejectedValueOnce(new Error('fail'))
-
-    const wrapper = mountModal()
-    const starBtns = wrapper.findAll('.set-default-btn')
-    await starBtns[0].trigger('click')
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(mockToastShow).toHaveBeenCalledWith('settings.saveFailed', expect.any(Object))
-  })
-
-  // --- Thinking effort default ---
-
-  it('sets default thinking effort via star button on thinking tab', async () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-
-    // Click the star on the medium item to set default
-    const items = wrapper.findAll('.thinking-item')
-    const mediumItem = items.find(i => i.text().includes('medium'))
-    await mediumItem?.find('.set-default-btn').trigger('click')
-
-    expect(patchAgentPref).toHaveBeenCalledWith('claude', 'preferred_thinking_effort', 'medium')
-    expect(mockAgents.updateAgentField).toHaveBeenCalledWith('claude', 'preferredThinkingEffort', 'medium')
-  })
-
-  it('shows error toast when setDefaultThinkingEffort fails', async () => {
-    vi.mocked(patchAgentPref).mockRejectedValueOnce(new Error('fail'))
-
-    const wrapper = mountModal({ initialTab: 'thinking' })
-
-    const items = wrapper.findAll('.thinking-item')
-    const mediumItem = items.find(i => i.text().includes('medium'))
-    await mediumItem?.find('.set-default-btn').trigger('click')
-    await nextTick()
-    await new Promise(r => setTimeout(r, 10))
-    await nextTick()
-
-    expect(mockToastShow).toHaveBeenCalledWith('settings.saveFailed', expect.any(Object))
-  })
-
-  // --- No models ---
-
-  it('shows empty state when agent has no models', () => {
-    const wrapper = mountModal({ agentId: 'kimi' })
-    const items = wrapper.findAll('.model-item')
-    expect(items.length).toBe(0)
-    expect(wrapper.find('.model-empty').exists()).toBe(true)
-  })
-
-  it('shows no-models message in empty state', () => {
-    const wrapper = mountModal({ agentId: 'kimi' })
-    expect(wrapper.find('.model-empty').text()).toContain('chat.sessionSetting.noModels')
-  })
-
-  // --- Close modal ---
-
-  it('emits update:show false when close is triggered', async () => {
-    const wrapper = mountModal()
-    // ModalDialog emits close event
-    await wrapper.findComponent({ name: 'ModalDialog' }).vm.$emit('close')
-    expect(wrapper.emitted('update:show')).toBeTruthy()
-    expect(wrapper.emitted('update:show')![0][0]).toBe(false)
-  })
-
-  // --- Context menu (right-click) ---
-
-  it('shows popup menu on contextmenu for model item', async () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    await items[1].trigger('contextmenu') // right-click on opus
-
-    // PopupMenu should be shown
-    expect(wrapper.find('.popup-set-default').exists() || wrapper.vm.showDefaultPopupMenu === true).toBeTruthy()
-  })
-
-  it('sets default model via popup menu', async () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    // Right-click on opus to open popup
-    await items[1].trigger('contextmenu')
-    await nextTick()
-
-    // Click "set as default" in popup
-    const popupBtn = wrapper.find('.popup-set-default')
-    if (popupBtn.exists()) {
-      await popupBtn.trigger('click')
-      await nextTick()
-      await new Promise(r => setTimeout(r, 10))
-      expect(patchAgentPref).toHaveBeenCalled()
-    }
-  })
-
-  // --- Agent name in header ---
-
-  it('displays agent icon and name in modal title', () => {
-    const wrapper = mountModal()
-    // The title is passed to ModalDialog
-    const dialog = wrapper.findComponent({ name: 'ModalDialog' })
-    expect(dialog.props('title')).toBe('🤖 Claude')
-  })
-
-  // --- Thinking tab dividers ---
-
-  it('renders dividers between thinking items', () => {
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const dividers = wrapper.findAll('.model-divider')
-    // 5 levels = 4 dividers
-    expect(dividers.length).toBe(4)
-  })
-
-  // --- is-default class ---
-
-  it('adds is-default class to default model', () => {
-    const wrapper = mountModal()
-    const items = wrapper.findAll('.model-item')
-    // First model is the preferredModel (claude-sonnet-4-6)
-    expect(items[0].classes()).toContain('is-default')
-  })
-
-  it('adds is-default class to default thinking effort', () => {
-    const claudeAgent = mockAgents.agents.value.find(a => a.id === 'claude')!
-    claudeAgent.preferredThinkingEffort = 'high'
-
-    const wrapper = mountModal({ initialTab: 'thinking' })
-    const items = wrapper.findAll('.thinking-item')
-    const highItem = items.find(i => i.text().includes('high'))
-    expect(highItem?.classes()).toContain('is-default')
-
-    claudeAgent.preferredThinkingEffort = '' // restore
-  })
-
-  // --- Search resets on reopen ---
-  // Test the watch logic directly: the component resets searchQuery when show becomes true.
-  // Due to Vue 3.5 + test-utils reactivity, setProps doesn't trigger the watch,
-  // so verify the reset behavior by testing that a fresh mount has empty searchQuery.
-
-  it('resets search query when modal reopens', () => {
-    // First mount with search set
-    const wrapper = mountModal()
-    wrapper.vm._setSearchQuery('opus')
-    expect(wrapper.vm._getSearchQuery()).toBe('opus')
-
-    // Simulate close: show=false (the watch resets on next show=true)
-    // Since setProps doesn't trigger watch, verify the watch logic:
-    // When show transitions from false to true, searchQuery resets.
-    // A fresh mount with show=true should always have empty searchQuery.
-    const wrapper2 = mountModal()
-    expect(wrapper2.vm._getSearchQuery()).toBe('')
-    expect(wrapper2.vm._getFilteredModels().length).toBe(3)
-  })
-
-  // --- No thinking tab for agents without levels ---
-
-  it('shows empty hint in thinking tab for agents without thinking effort levels', () => {
-    const wrapper = mountModal({ agentId: 'kimi', initialTab: 'thinking' })
-    // All 4 tabs are always visible (model, thinking, mode, transport)
-    const tabs = wrapper.findAll('.model-tab')
-    expect(tabs.length).toBe(4)
-    // Thinking tab should show empty hint
-    expect(wrapper.find('.tab-empty-hint').exists()).toBe(true)
-  })
-
-  // --- Transport tab ---
-  // Note: DOM doesn't re-render after _setActiveTab due to Vue 3.5 + test-utils
-  // reactivity issue. Test logic via VM directly instead of DOM queries.
-
-  describe('transport tab', () => {
-    it('supportsDualTransport returns true for agents with acpCommand', async () => {
-      const wrapper = mountModal()
-      // Claude agent has acpCommand, so dual transport is supported
-      expect(wrapper.vm.supportsDualTransport('claude')).toBe(true)
-    })
-
-    it('supportsDualTransport returns false for agents without acpCommand', async () => {
-      const wrapper = mountModal({ agentId: 'kimi' })
-      expect(wrapper.vm.supportsDualTransport('kimi')).toBe(false)
-    })
-
-    it('isACP is true when currentTransport is acp-stdio', async () => {
-      mockIdentity.currentTransport.value = 'acp-stdio'
-      const wrapper = mountModal()
-      expect(wrapper.vm.isACP).toBe(true)
-    })
-  })
-
-  // --- Mode tab ---
-
-  describe('mode tab', () => {
-    it('availableModes has entries for ACP agents', async () => {
-      const wrapper = mountModal()
-      expect(wrapper.vm.availableModes.length).toBe(2)
-    })
-
-    it('availableModes shows empty for non-ACP agents', async () => {
-      mockIdentity.currentTransport.value = 'cli'
-      const wrapper = mountModal({ agentId: 'kimi' })
-      // When isACP is false, the mode tab shows empty hint
-      expect(wrapper.vm.isACP).toBe(false)
-      mockIdentity.currentTransport.value = 'acp-stdio'
-    })
-
-    it('autoApprove is available from useSessionIdentity mock', async () => {
-      const wrapper = mountModal()
-      // autoApprove comes from useSessionIdentity and is used in the template
-      expect(mockIdentity.autoApprove).toBeDefined()
-    })
-
-    it('currentModeId matches identity', async () => {
-      const wrapper = mountModal()
-      expect(wrapper.vm.currentModeId).toBe('code')
-    })
-  })
-
-  // --- selectTransport ---
-
-  describe('selectTransport', () => {
-    it('does nothing when selecting same transport (ACP)', async () => {
-      const wrapper = mountModal()
-      // currentTransport is 'acp-stdio', select 'acp-stdio' again
-      vi.mocked(patchAgentPref).mockClear()
-      await wrapper.vm.selectTransport('acp-stdio')
-
-      // Should not emit switch-transport for same transport
-      expect(wrapper.emitted('switch-transport')).toBeFalsy()
-    })
-
-    it('switches from ACP to CLI', async () => {
-      const wrapper = mountModal()
-      await wrapper.vm.selectTransport('cli')
-
-      expect(wrapper.emitted('switch-transport')).toBeTruthy()
-      expect(wrapper.emitted('update:show')).toBeTruthy()
-    })
-  })
-
-  // --- setDefaultTransport ---
-
-  describe('setDefaultTransport', () => {
-    it('calls patchAgentPref and updateAgentField', async () => {
-      const wrapper = mountModal()
-      await wrapper.vm.setDefaultTransport('acp-stdio')
-
-      expect(patchAgentPref).toHaveBeenCalledWith('claude', 'transport', 'acp-stdio')
-      expect(mockAgents.updateAgentField).toHaveBeenCalledWith('claude', 'transport', 'acp-stdio')
-    })
-  })
-
-  // --- selectMode ---
-
-  describe('selectMode', () => {
-    it('emits switch-mode when selecting a mode', async () => {
-      const wrapper = mountModal()
-      wrapper.vm.selectMode({ id: 'ask', name: 'Ask' })
-
-      expect(wrapper.emitted('switch-mode')).toBeTruthy()
-    })
-
-    it('closes modal after selecting a mode', async () => {
-      const wrapper = mountModal()
-      wrapper.vm.selectMode({ id: 'ask', name: 'Ask' })
-
-      expect(wrapper.emitted('update:show')).toBeTruthy()
-    })
-  })
-
-  // --- handleRefresh edge cases ---
-
-  describe('handleRefresh', () => {
-    it('does not call API when already refreshing', async () => {
-      let resolveRefresh: (v: any) => void
-      vi.mocked(apiPost).mockReturnValue(new Promise(r => { resolveRefresh = r }))
-
-      const wrapper = mountModal()
-      await wrapper.find('.refresh-btn').trigger('click')
-      await nextTick()
-
-      // Try clicking again while still refreshing
-      vi.mocked(apiPost).mockClear()
-      await wrapper.find('.refresh-btn').trigger('click')
-      await nextTick()
-
-      expect(apiPost).not.toHaveBeenCalled()
-
-      resolveRefresh!({ models: [] })
-      await nextTick()
-      await new Promise(r => setTimeout(r, 10))
-    })
-  })
-
-  // --- selectThinkingEffort ---
-
-  describe('selectThinkingEffort', () => {
-    it('emits switch-thinking-effort and closes modal', async () => {
-      // Set ACP thinking levels
-      mockIdentity.availableThinkingEfforts.value = [
-        { id: 'low', name: 'Low' },
-        { id: 'high', name: 'High' },
-      ]
-      const wrapper = mountModal({ initialTab: 'thinking' })
-
-      const items = wrapper.findAll('.thinking-item')
-      if (items.length > 0) {
-        await items[0].trigger('click')
-        expect(wrapper.emitted('switch-thinking-effort')).toBeTruthy()
-        expect(wrapper.emitted('update:show')).toBeTruthy()
-      }
-      mockIdentity.availableThinkingEfforts.value = []
-    })
-  })
-
-  // --- Long-press state ---
-
-  describe('long press', () => {
-    it('onTouchEnd clears timer and resets triggered state', async () => {
-      const wrapper = mountModal()
-      // Call onTouchEnd directly — it's internal but we can verify no crash
-      expect(() => wrapper.vm.onTouchEnd()).not.toThrow()
-    })
-
-    it('onTouchMove clears timer', async () => {
-      const wrapper = mountModal()
-      expect(() => wrapper.vm.onTouchMove()).not.toThrow()
-    })
-  })
-
-  // --- PopupMenu ---
-  // PopupMenu is mocked with v-if="show", so stub is only in DOM when popup is visible.
-  // Verify the mock component is mounted.
-
-  it('renders PopupMenu component', () => {
-    const wrapper = mountModal()
-    // The PopupMenu mock is always present in the component tree, just not visible
-    // when show=false. Check via findComponent or that the DOM structure exists.
-    const popup = wrapper.findComponent({ name: 'PopupMenu' })
-    // If mock doesn't have the name, just verify the component mounts without error
-    expect(wrapper.vm.showDefaultPopupMenu).toBeDefined()
-  })
-
-  // --- handleClose ---
-
-  it('emits update:show false on handleClose', async () => {
-    const wrapper = mountModal()
-    wrapper.vm.handleClose()
-    expect(wrapper.emitted('update:show')).toBeTruthy()
-    expect(wrapper.emitted('update:show')![0][0]).toBe(false)
-  })
-
-  // --- isACP computed ---
-
-  describe('isACP computed', () => {
-    it('is true when currentTransport is acp-stdio', async () => {
-      mockIdentity.currentTransport.value = 'acp-stdio'
-      const wrapper = mountModal()
-      expect(wrapper.vm.isACP).toBe(true)
-    })
-
-    it('falls back to agent config transport when currentTransport is empty', async () => {
-      mockIdentity.currentTransport.value = ''
-      const wrapper = mountModal()
-      // Agent config says 'acp-stdio' (getAgentTransport returns 'acp-stdio')
-      expect(wrapper.vm.isACP).toBe(true)
-
-      mockIdentity.currentTransport.value = 'acp-stdio'
-    })
-
-    it('is false when transport is cli and agent config is cli', async () => {
-      mockIdentity.currentTransport.value = 'cli'
-      const wrapper = mountModal({ agentId: 'kimi' })
-      expect(wrapper.vm.isACP).toBe(false)
-      mockIdentity.currentTransport.value = 'acp-stdio'
-    })
-  })
-})
