+++ web/src/components/chat/ChatInputBar.vue
@@ -132,12 +132,12 @@
           ⚙️ {{ t('chat.quickSend.edit') }}
         </button>
       </PopupMenu>
-      <!-- Session settings modal -->
-      <SessionSettingModal
-        :show="settingsDrawer.effectiveOpen.value"
+      <!-- Session settings drawer -->
+      <SessionDrawer
+        :open="settingsDrawer.effectiveOpen.value"
         :agent-id="currentAgentId"
-        :initial-tab="settingsModalInitialTab"
-        @update:show="$event ? settingsDrawer.open() : settingsDrawer.close()"
+        :initial-tab="settingsDrawerInitialTab"
+        @close="settingsDrawer.close()"
         @switch-model="handleSwitchModel"
         @switch-thinking-effort="handleSwitchThinkingEffort"
         @switch-mode="handleSwitchMode"
@@ -202,18 +202,18 @@
     </div>
     <!-- Session info bar (model + mode + thinking + transport) -->
     <div class="chat-session-info" v-if="currentModelName || showModeInfo || showThinkingInfo || showTransportInfo || showUsageInfo">
-      <span class="session-info-model" @click.stop="openSettingsModal('model')"><Cpu :size="11" />{{ currentModelName }}</span>
+      <span class="session-info-model" @click.stop="openSettingsDrawer('model')"><Cpu :size="11" />{{ currentModelName }}</span>
       <template v-if="showModeInfo">
         <span class="session-info-divider"></span>
         <span class="session-info-mode" :class="{ 'session-info-mode-auto': autoApprove }" @click.stop="onModeClick" v-long-press="onModeLongPress" @mousedown.stop="onModeMouseDown" @mouseup.stop="onModeMouseUp"><Compass :size="11" />{{ currentModeName }}</span>
       </template>
       <template v-if="showThinkingInfo">
         <span class="session-info-divider"></span>
-        <span class="session-info-thinking" @click.stop="openSettingsModal('thinking')"><Brain :size="11" />{{ currentThinkingEffortName }}</span>
+        <span class="session-info-thinking" @click.stop="openSettingsDrawer('thinking')"><Brain :size="11" />{{ currentThinkingEffortName }}</span>
       </template>
       <template v-if="showTransportInfo">
         <span class="session-info-divider"></span>
-        <span class="session-info-transport" @click.stop="openSettingsModal('transport')"><Cable :size="11" />{{ currentTransport === 'acp-stdio' ? 'ACP' : 'CLI' }}</span>
+        <span class="session-info-transport" @click.stop="openSettingsDrawer('transport')"><Cable :size="11" />{{ currentTransport === 'acp-stdio' ? 'ACP' : 'CLI' }}</span>
       </template>
       <template v-if="showUsageInfo">
         <span class="session-info-divider"></span>
@@ -242,7 +242,7 @@ import PopupMenu from '@/components/common/PopupMenu.vue'
 import AttachDrawer from '@/components/chat/AttachDrawer.vue'
 import { useTabDrawer } from '@/composables/useTabDrawer'
 import QuickSendDrawer from '@/components/chat/QuickSendDrawer.vue'
-import SessionSettingModal from '@/components/chat/SessionSettingModal.vue'
+import SessionDrawer from '@/components/chat/SessionDrawer.vue'
 import { createStopButtonMachine } from '@/utils/stopButtonMachine.ts'
 import { useDialog } from '@/composables/useDialog.ts'
 import { useQuickSend } from '@/composables/useQuickSend'
@@ -278,7 +278,7 @@ function onModeClick() {
     modeMouseLongFired = false
     return
   }
-  openSettingsModal('mode')
+  openSettingsDrawer('mode')
 }
 
 // Long-press on mode chip → toggle auto-approve
@@ -324,7 +324,7 @@ const usageColor = computed(() => {
 const dialog = useDialog()
 const quickSendStore = useQuickSend()
 const { items: quickSendItems, fetchItems } = quickSendStore
-const settingsModalInitialTab = ref('model')
+const settingsDrawerInitialTab = ref('model')
 const quickSendDrawer = useTabDrawer('chat', quickSendStore.showEditDialog)
 const settingsDrawer = useTabDrawer('chat')
 
@@ -426,8 +426,8 @@ const attachMenuRef = ref(null) // kept for ref stability, no longer used for Po
 const showQuickMenu = ref(false)
 const sendBtnRef = ref(null)
 
-function openSettingsModal(tab) {
-  settingsModalInitialTab.value = tab
+function openSettingsDrawer(tab) {
+  settingsDrawerInitialTab.value = tab
   settingsDrawer.open()
 }
